requirements.txt
aiocron==1.8
aiofiles==0.6.0
aiohttp
aiohttp-cors==0.7.0
aiomysql==0.0.21
aiosqlite==0.17.0
appdirs==1.4.4
arrow==0.17.0
astunparse==1.6.3
async-cse==0.3.0
async-generator==1.10
async-timeout==3.0.1
asyncio-extras==1.3.2
asyncpraw==7.3.0
asyncprawcore==2.2.0
attrs==21.2.0
beautifulsoup4==4.9.3
betterstatcord.py==3.4.1
bleach==3.3.0
braceexpand==0.1.7
button-paginator @ git+https://github.com/antinuke0day/Button_paginator
certifi==2021.5.30
cffi==1.15.0
cfgv==3.3.0
chardet==4.0.0
colorama==0.4.4
coloredlogs==14.3
colorgram.py==1.2.0
croniter==1.3.4
cycler==0.10.0
distlib==0.3.2
durations-nlp==1.0.1
filelock==3.0.12
geographiclib==1.52
geopy==2.2.0
humanfriendly==9.2
humanize==3.9.0
identify==2.2.10
idna==2.10
import-expression==1.1.4
jishaku
kdtree==0.16
kiwisolver==1.3.1
asyncmy
asyncpg
matplotlib==3.4.2
multidict==5.1.0
nekos.py==1.0.3
nodeenv==1.6.0
numpy==1.20.3
oauthlib==3.1.1
packaging==20.9
parsedatetime==2.6
Pillow==8.2.0
pre-commit==2.13.0
psutil==5.8.0
PyMySQL==0.9.3
pyparsing==2.4.7
pyreadline==2.1
PySocks==1.7.1
python-dateutil==2.8.1
python-dotenv==0.15.0
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
PyYAML==5.4.1
random-user-agent==1.0.1
regex==2021.4.4
requests==2.25.1
requests-oauthlib==1.3.0
schedule==1.1.0
scipy==1.6.1
six==1.16.0
soupsieve==2.2.1
timeago==1.0.15
timezonefinder==5.2.0
toml==0.10.2
instagram_private_api
instaloader
tweepy==3.10.0
TwitterAPI==2.7.12
typing-extensions==3.10.0.0
tzdata==2022.1
tzlocal==4.2
update-checker==0.18.0
urllib3==1.26.5
virtualenv==20.4.7
webencodings==0.5.1
wikipedia==1.4.0
wordcloud==1.8.1
yarl==1.6.3
zope.interface==5.4.0
